Fixed JSON error preventing login
TTR recently added a new 'ip' field to the login response. It is set to null
since it was just added and is unused so far, but this null value caused
JSONObject.getString() to throw an exception and prevent login. So, I added
some safety checking, which passes on the null value to the returned Map,
and leaves a warning in the log.